updated version of Nick using Alpine and made working for X86 by Alex

+++ proxy-joplin-editor.sh
@@ -16,6 +16,17 @@ cd "$(dirname "$0")";
 . ./config-defaults.sh
 cd ${CURR_WD}
 
+if [[ ! -f $TEMP_APPEND_FILE ]]; then
+    echo "File with content to append not found"
+    exit 1
+fi
+
+# Append 2 x newlines ahead of new content.
+# This will prevent conflicts with the note title which takes up the first two lines
+echo >> $1
+echo >> $1
+
+# Append the content of our reserved file to the existing file, location of which was provided by Joplin
 cat $TEMP_APPEND_FILE >> $1
 
 rm $TEMP_APPEND_FILE
